requirements.txt
annotated-types==0.7.0
anyio==4.6.0
fastapi==0.115.0
greenlet==3.1.1
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1  # Для импорта переменных среды
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2